posts/super_bowl/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
folium>=0.14.0
matplotlib>=3.7.0

# BigQuery (optional - uses cached data by default)
google-cloud-bigquery>=3.10.0
google-auth>=2.16.0
db-dtypes>=1.1.0

# Data processing
pyarrow>=12.0.0
